Broke down servers into their own classes
